pyproject.toml
[project]
name = "uav-service"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
license = "Proprietary"
packages = [
    {include = "uav_service", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.120.4"}
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = {extras = ["mypy"], version = "^2.11.9"}
pydantic-settings = "^2.11.0"
numpy = "^2.3.4"
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "<4.1"

[tool.poetry.group.lint.dependencies]
mypy = "^1.18.2"
isort = "^6.1.0"
black = "^25.9.0"
flake8 = "^7.3.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
